requirements.txt
GitPython==3.1.40
PyYAML==6.0.1
